pyproject.toml
[project]
name = "langchain-v1-test"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "concurrent-log-handler==0.9.28",
    "langchain==1.2.1",
    "langchain-chroma==1.1.0",
    "langchain-community==0.4.1",
    "langchain-huggingface>=1.2.0",
    "langchain-openai==1.1.6",
    "langchain-text-splitters==1.1.0",
    "langgraph-checkpoint-postgres==3.0.4",
    "psycopg[binary]>=3.3.3",
    "pypdf==6.6.0",
    "sentence-transformers>=5.2.3",
    "socksio==1.0.0",
]
